requirements.txt
requests
pandas
pyarrow
python-dotenv
pytest
jupyterlab
matplotlib

